moved sensitive check to members api
- emails need to send from the server anyway
